Error handling in start and stop scripts

+++ k8s/scripts/eess-k8s-stop.sh
@@ -12,9 +12,14 @@ NAMESPACE="eess-k8s"
 
 shopt -s nullglob
 
+# Raise error if minikube is not running
+if ! (minikube status --format '{{.Host}}' 2>/dev/null | grep -q "Running"); then
+    echo "Minikube is not running. Please start the cluster before running this script."
+    exit 1
+fi
+
 # Delete the namespace and all resources within it. This will also delete services, deployments, and ConfigMaps.
 kubectl delete namespace "$NAMESPACE" --ignore-not-found
-echo "Namespace $NAMESPACE deleted successfully and all resources within it."
 
 if [ "$1" == "-M" ]; then
     # Stop the cluster with minikube if it's running.
@@ -25,4 +30,3 @@ if [ "$1" == "-M" ]; then
         echo "Minikube is not running."
     fi
 fi
-echo "Kubernetes cluster stopped and services, deployments, and ConfigMaps deleted successfully."
